Store all task results, even if some files don't exist.
When storing the results of a task, STARS should grab as many files as
possible. However, with the previous ordering of statements, if storing
a result failed, all subsequent results would be skipped (since
`success` was false and `and` returns the first false).

This patch reverses the order of statements such that an attempt will be
made to fetch each result and success will return true only if all
results are successfully stored.