Change lookup for `env.sh`.
When looking for `env.sh`, first look in the directory the effective
directory of the script (if the script is launched via a symlink, the
effective directory is the directory of the symlink. Otherwise the
effective directory is the actual directory of the script).

If the file does not exist there, look in the actual directory of the
script.

This allows scripts to be invoked even when the `pwd` is not the script
directory. Also, multiple clusters can be controlled by a single
instance of the scripts by creating symlinks of each script in a
directory with a different `env.sh` file.