Silence errors related to `stop` and `stopnow`.
`stop` and `stopnow` are client scripts, not deployment scripts. This
was changed at some point before this code was committed to git.
Unfortunately, the `package script` was not updated and, as a result, it
generates error messages while running.

This patch removes these files from the section of the `package` script
where the deployment package is built.