Fix `start`ing of jobs on a remote management node.
Workfiles submitted to a remote management node were initially copied
with a `.tmp` suffix and then the `.tmp` suffix is removed. However,
`workfilehandler.py` does not ignore files with the `.tmp` suffix; it
treats them as a new job and starts processing them immediately. Thus,
once the file is renamed, the management node will detect it as a new
work file and start processing it again. This results in (at least
partial) duplication of submitted workfiles.

Since `workfilehandler.py` ignores hidden files (i.e., files with a
leading `.` in their name), this patch changes job submission to
initially submit workfiles as hidden files and then renames them without
the leading `.`.